Fixed typo in DataTypes of ODE functions args

+++ SIMULATIONS/HD/WD/user_param.h
@@ -27,9 +27,9 @@
 #define MB 2 //barionic mass
 
 int buildModel();
-float  interior_ode(double x, double y[], int i);
-float  taylor_ode(double x, double y[], int i);
+float  interior_ode(float x, float y[], int i);
+float  taylor_ode(float x, float y[], int i);
 
-float eos(double valor, int i);
+float eos(float valor, int i);
 
 float massDensity, polytropicExp, polytropicK;
